+++ include/galsim/Angle.h
@@ -188,13 +188,13 @@ namespace galsim {
         friend std::ostream& operator<<(std::ostream& os, Angle theta)
         { os << theta._val; return os; }
 
-        /// Wraps this angle to the range (-pi, pi]
-        Angle wrap() {
+        /// Wraps this angle to the range (c-pi, c+pi]
+        Angle wrap(double center=0.) {
             const double TWOPI = 2.*M_PI;
-            double new_val = std::fmod(_val, TWOPI); // now in range (-TWOPI, TWOPI)
+            double new_val = std::fmod(_val-center, TWOPI); // now in range (-TWOPI, TWOPI)
             if (new_val <= -M_PI) new_val += TWOPI;
             if (new_val > M_PI) new_val -= TWOPI;
-            return Angle(new_val, radians);
+            return Angle(new_val+center, radians);
         }
 
         double sin() const { return std::sin(_val); }